+++ example/1-pipeline/serverless/rand.go
@@ -6,13 +6,13 @@ import (
 )
 
 // Handler will handle the raw data.
-func Handler(data []byte) (byte, []byte) {
+func Handler(data []byte) (uint32, []byte) {
 	randint := binary.LittleEndian.Uint32(data)
 	log.Printf("Generate random uint32: %d (%# x)", randint, data)
 	return 0, nil
 }
 
 // DataTags describes the type of data this serverless function observed.
-func DataTags() []byte {
-	return []byte{0x01}
+func DataTags() []uint32 {
+	return []uint32{0x01}
 }
