+++ example/2-iopipe/serverless/counter.go
@@ -5,7 +5,7 @@ import (
 )
 
 // Handler counts how many bytes received
-func Handler(data []byte) (byte, []byte) {
+func Handler(data []byte) (uint32, []byte) {
 	log.Printf("Got: %d", len(data))
 
 	// return 0, nil will tell zipper end the workflow.
@@ -13,6 +13,6 @@ func Handler(data []byte) (byte, []byte) {
 }
 
 // DataTags describes the type of data this serverless function observed.
-func DataTags() []byte {
-	return []byte{0x01}
+func DataTags() []uint32 {
+	return []uint32{0x01}
 }
