+++ example/3-multi-sfn/stream-fn-2/app.go
@@ -48,7 +48,7 @@ func main() {
 	select {}
 }
 
-func handler(data []byte) (byte, []byte) {
+func handler(data []byte) (uint32, []byte) {
 	v := Float32frombytes(data)
 	result, err := computePeek(context.Background(), v)
 	if err != nil {
