+++ example/5-backflow/sfn-2/main.go
@@ -38,7 +38,7 @@ func main() {
 	select {}
 }
 
-func handler(data []byte) (byte, []byte) {
+func handler(data []byte) (uint32, []byte) {
 	// got
 	noise, err := strconv.Atoi(string(data))
 	if err != nil {
